(BE/User)Fix: jwtUtil의 refreshToken 삭제 코드 수정

+++ back/groogroo/src/main/java/com/jwcg/groogroo/util/JwtUtil.java
@@ -214,10 +214,10 @@ public class JwtUtil {
      */
     @Transactional
     public void deleteRefreshToken(String accessToken) {
-        RefreshToken token = tokenRepository.findByAccessToken(accessToken)
-                .orElseThrow(IllegalArgumentException::new);
-
-        tokenRepository.delete(token);
+        RefreshToken token = tokenRepository.findByAccessToken(accessToken).get();
+        if(token != null){
+            tokenRepository.delete(token);
+        }
     }
